fixing wmt pytypes
