+++ algorithmic_efficiency/workloads/wmt/workload.py
@@ -5,7 +5,6 @@ import jax
 import numpy as np
 import torch
 
-from algorithmic_efficiency import param_utils
 from algorithmic_efficiency import spec
 from algorithmic_efficiency.workloads.wmt import decode
 from algorithmic_efficiency.workloads.wmt import input_pipeline
@@ -181,16 +180,6 @@ class BaseWmtWorkload(spec.Workload):
           'This should not happen, workload.init_model_fn() should be called '
           'before workload.param_shapes!')
     return self._param_shapes
-
-  @property
-  def model_params_types(self):
-    if self._param_shapes is None:
-      raise ValueError(
-          'This should not happen, workload.init_model_fn() should be called '
-          'before workload.param_shapes!')
-    if self._param_types is None:
-      self._param_types = param_utils.jax_param_types(self._param_shapes)
-    return self._param_types
 
   def output_activation_fn(self,
                            logits_batch: spec.Tensor,
